server/requirements.txt
boto3
awscli
flask 
flask-cors
upstash-vector
sentence-transformers
beautifulsoup4
requests
numpy
python-dotenv
unidecode